+++ devtools/replidev/src/command/mod.rs
@@ -1,2 +1,3 @@
+pub mod certs;
 pub mod conf;
 pub mod deps;